Experimental lookup for partial postcodes
Very much WIP, to do with a paired `frontend` branch, which has a
rake task to compare accuracy for a full and partial postcode,
when matching against local transactions.

The partial postcode API in MapIt doesn't return areas, but these
can be looked up from the point API using the co-ords returned for
the partial.

Rather than making callers of GDS API need to be aware of this we
can mask the difference by combining the two MapIt responses:
 - the partial API gives the standard lat/long/etc parts
 - the point API gives a set of areas

I wonder if this is something MapIt itself could do, seems like a
reasonable use case. But for now this allows us to test partial
postcode area lookup with a similar interface to the existing
postcode lookup